feat: add pool api

+++ src/SchrodingerServer.Application.Contracts/Dtos/Cat/CombineOutput.cs
@@ -6,6 +6,6 @@ namespace SchrodingerServer.Dtos.Cat;
 public class CombineOutput
 {
     public List<string> AdoptIds { get; set; }
-    public int Level { get; set; }
+    public long Level { get; set; }
     public ByteString Signature { get; set; } 
 }
